backend/pyproject.toml
[project]
name = "multi-tenant-saas-backend"
version = "0.1.0"
description = "Multi-tenant SaaS platform with AI capabilities"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "structlog>=24.4.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "coverage>=7.6.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["B008", "UP046"]  # B008: Depends() in defaults is standard FastAPI pattern

[tool.black]
line-length = 99
target-version = ["py312"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
